AVFilter, examples
